+++ assets/javascripts/kitten/components/form/autocomplete/autocomplete.stories.js
@@ -166,7 +166,7 @@ const args = {
   onKeyDown: () => {},
   onSelect: () => {},
   isLoading: false,
-  variant: 'andromeda',
+  variant: 'orion',
 }
 
 const argTypes = {