+++ assets/javascripts/kitten/components/form/checkbox-set/checkbox-set.stories.js
@@ -37,7 +37,7 @@ export default {
       },
     ],
     disabled: false,
-    variant: 'andromeda',
+    variant: 'orion',
   },
   argTypes: {
     name: {