+++ assets/javascripts/kitten/components/organisms/header-nav/stories/kisskissbankbank-new.js
@@ -214,9 +214,8 @@ const SearchInput = () => {
 
       <Button
         size="tiny"
-        variant="orion"
         type="submit"
-        fluid
+        fit="fluid"
         className="k-u-margin-bottom-double k-u-margin-horizontal-double"
         style={{
           width: 'calc(100% - 40px)',