+++ assets/javascripts/kitten/components/organisms/vertical-stepper/components/title.js
@@ -43,5 +43,5 @@ Title.protoTypes = {
 }
 
 Title.defaultProps = {
-  variant: 'andromeda',
+  variant: 'orion',
 }