+++ src/components/form/date-picker/index.js
@@ -50,21 +50,20 @@ var cellSize = (0, _typography.pxToRem)(50);
 var tinyCellSize = (0, _typography.pxToRem)(40);
 var dayPickerPadding = (0, _typography.pxToRem)(30);
 var tinyDayPickerPadding = (0, _typography.pxToRem)(20);
-var fontSize = (0, _styledComponents.css)(["font-size:", "rem;"], _typographyConfig.default.fontSize / _typographyConfig.default.scaleMultiplier);
 
 var StyledDatePicker = _styledComponents.default.div.withConfig({
   displayName: "date-picker__StyledDatePicker",
   componentId: "sc-1jr1agu-0"
-})(["box-sizing:border-box;.DayPickerInput-Overlay{box-shadow:none;margin-top:18px;min-width:calc(7 * ", " + 2 * ", ");@media (min-width:", "px){min-width:calc(7 * ", " + 2 * ", ");}&:before{content:'';width:0;height:0;border-left:", " solid transparent;border-right:", " solid transparent;", " position:absolute;top:-", ";left:", ";}}.DayPicker-Caption{text-align:center;height:", ";margin:-", " -", " 0;overflow:hidden;@media (min-width:", "px){margin:-", " -", " 0;}", " & div{padding:", " 0;margin:0 ", ";", "}}.DayPicker-Weekday{", " ", "}.DayPicker-Weekday,.DayPicker-Day{padding:0;width:", ";height:", ";vertical-align:middle;box-sizing:border-box;", " @media (min-width:", "px){width:", ";height:", ";}}.DayPicker-Day{", " border-radius:0;&.DayPicker-Day--outside{border:0;}&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}}.DayPicker-Day--today{", " ", "}.DayPicker-Day--selected{&:not(.DayPicker-Day--disabled){&:not(.DayPicker-Day--outside){", " &:hover{", "}}}}.DayPicker-Day--disabled{", "}.DayPicker-Month{margin:0;}.DayPicker-wrapper{padding:0 ", " ", " ", ";@media (min-width:", "px){padding:0 ", " ", " ", ";}&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}}.DayPicker{", " ", " ", " &:not(.DayPicker--interactionDisabled){.DayPicker-Day{&:not(.DayPicker-Day--disabled){&:not(.DayPicker-Day--selected){&:not(.DayPicker-Day--outside){&:hover{", " ", "}}}}}}}"], tinyCellSize, tinyDayPickerPadding, _screenConfig.ScreenConfig.S.min, cellSize, dayPickerPadding, (0, _typography.pxToRem)(9), (0, _typography.pxToRem)(9), function (_ref) {
+})(["box-sizing:border-box;.DayPickerInput-Overlay{box-shadow:none;margin-top:18px;min-width:calc(7 * ", " + 2 * ", ");@media (min-width:", "px){min-width:calc(7 * ", " + 2 * ", ");}&:before{content:'';width:0;height:0;border-left:", " solid transparent;border-right:", " solid transparent;", " position:absolute;top:-", ";left:", ";}}.DayPicker-Caption{text-align:center;height:", ";margin:-", " -", " 0;overflow:hidden;@media (min-width:", "px){margin:-", " -", " 0;}", " & div{padding:", " 0;margin:0 ", ";font-size:", ";}}.DayPicker-Weekday{", " ", "}.DayPicker-Weekday,.DayPicker-Day{padding:0;width:", ";height:", ";vertical-align:middle;box-sizing:border-box;font-size:", ";@media (min-width:", "px){width:", ";height:", ";}}.DayPicker-Day{", " border-radius:0;&.DayPicker-Day--outside{border:0;}&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}}.DayPicker-Day--today{", " ", "}.DayPicker-Day--selected{&:not(.DayPicker-Day--disabled){&:not(.DayPicker-Day--outside){", " &:hover{", "}}}}.DayPicker-Day--disabled{", "}.DayPicker-Month{margin:0;}.DayPicker-wrapper{padding:0 ", " ", " ", ";@media (min-width:", "px){padding:0 ", " ", " ", ";}&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus:not(:focus-visible){outline-color:transparent;}&:focus-visible{outline-color:", ";}}.DayPicker{font-size:", ";", " ", " &:not(.DayPicker--interactionDisabled){.DayPicker-Day{&:not(.DayPicker-Day--disabled){&:not(.DayPicker-Day--selected){&:not(.DayPicker-Day--outside){&:hover{", " ", "}}}}}}}"], tinyCellSize, tinyDayPickerPadding, _screenConfig.ScreenConfig.S.min, cellSize, dayPickerPadding, (0, _typography.pxToRem)(9), (0, _typography.pxToRem)(9), function (_ref) {
   var styles = _ref.styles;
   return (0, _styledComponents.css)(["border-bottom:", " solid ", ";"], (0, _typography.pxToRem)(8), styles.header.backgroundColor);
 }, (0, _typography.pxToRem)(8), (0, _typography.pxToRem)(15), (0, _typography.pxToRem)(70), (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(22), _screenConfig.ScreenConfig.S.min, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(32), function (_ref2) {
   var styles = _ref2.styles;
   return (0, _styledComponents.css)(["background-color ", ";color:", ";"], styles.header.backgroundColor, styles.header.color);
-}, (0, _typography.pxToRem)(23), (0, _typography.pxToRem)(63), fontSize, _typographyConfig.default.fontStyles.regular, function (_ref3) {
+}, (0, _typography.pxToRem)(23), (0, _typography.pxToRem)(63), (0, _typography.stepToRem)(1), _typographyConfig.default.fontStyles.regular, function (_ref3) {
   var styles = _ref3.styles;
   return (0, _styledComponents.css)(["color:", ";"], styles.weekdaysColor);
-}, tinyCellSize, tinyCellSize, fontSize, _screenConfig.ScreenConfig.S.min, cellSize, cellSize, function (_ref4) {
+}, tinyCellSize, tinyCellSize, (0, _typography.stepToRem)(1), _screenConfig.ScreenConfig.S.min, cellSize, cellSize, function (_ref4) {
   var styles = _ref4.styles;
   return (0, _styledComponents.css)(["border:solid ", " ", ";"], styles.borderColor, borderSize);
 }, _colorsConfig.default.primary4, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(2), _colorsConfig.default.primary4, _typographyConfig.default.fontStyles.regular, function (_ref5) {
@@ -79,7 +78,7 @@ var StyledDatePicker = _styledComponents.default.div.withConfig({
 }, function (_ref8) {
   var styles = _ref8.styles;
   return (0, _styledComponents.css)(["color:", ";"], styles.day.disabled.color);
-}, tinyDayPickerPadding, tinyDayPickerPadding, tinyDayPickerPadding, _screenConfig.ScreenConfig.S.min, dayPickerPadding, dayPickerPadding, dayPickerPadding, _colorsConfig.default.primary4, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(2), _colorsConfig.default.primary4, fontSize, _typographyConfig.default.fontStyles.light, function (_ref9) {
+}, tinyDayPickerPadding, tinyDayPickerPadding, tinyDayPickerPadding, _screenConfig.ScreenConfig.S.min, dayPickerPadding, dayPickerPadding, dayPickerPadding, _colorsConfig.default.primary4, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(2), _colorsConfig.default.primary4, (0, _typography.stepToRem)(1), _typographyConfig.default.fontStyles.light, function (_ref9) {
   var styles = _ref9.styles;
   return (0, _styledComponents.css)(["border:solid ", " ", ";"], styles.borderColor, borderSize);
 }, _typographyConfig.default.fontStyles.light, function (_ref10) {