+++ src/esm/constants/typography-config.js
@@ -6,7 +6,6 @@ var TYPOGRAPHY = {
   fontWeight: 400,
   lineHeight: 1.5,
   // rem
-  scaleMultiplier: 1.125,
   fontStyles: {
     light: {
       fontFamily: 'Maax, Helvetica, Arial, sans-serif',